# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Travel-Api Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/travels/{travelSlug}/tours": {
            "get": {
                "tags": [
                    "Tour"
                ],
                "summary": "Get paginate list of tours",
                "description": "Returns list of tours",
                "operationId": "getListTours",
                "parameters": [
                    {
                        "name": "travelSlug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "priceFrom",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "priceTo",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "dateFrom",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "dateTo",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "price"
                        }
                    },
                    {
                        "name": "sortOrder",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "asc"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TourResource"
                                },
                                "example": {
                                    "data": [
                                        {
                                            "id": "99eb9c87-79ab-491d-9bb7-a29e77e14e79",
                                            "name": "Magnam totam illum.",
                                            "starting_date": "2023-08-18",
                                            "ending_date": "2023-08-23",
                                            "price": 154.6
                                        },
                                        {
                                            "id": "2",
                                            "name": "Magnam totam illum.",
                                            "starting_date": "2023-08-18",
                                            "ending_date": "2023-08-23",
                                            "price": 154.6
                                        }
                                    ],
                                    "links": {
                                        "first": "/api/v1/travels/facilis-fugit-iusto/tours?page=1",
                                        "last": "/api/v1/travels/facilis-fugit-iusto/tours?page=1",
                                        "prev": null,
                                        "next": null
                                    },
                                    "meta": {
                                        "current_page": 1,
                                        "from": 1,
                                        "last_page": 1,
                                        "links": [
                                            {
                                                "url": null,
                                                "label": "&laquo; Previous",
                                                "active": false
                                            },
                                            {
                                                "url": "/api/v1/travels/facilis-fugit-iusto/tours?page=1",
                                                "label": "1",
                                                "active": true
                                            },
                                            {
                                                "url": null,
                                                "label": "Next &raquo;",
                                                "active": false
                                            }
                                        ],
                                        "path": "/api/v1/travels/facilis-fugit-iusto/tours",
                                        "per_page": 15,
                                        "to": 4,
                                        "total": 4
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Tour": {
                "title": "Tour",
                "description": "Tour model",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "UUid",
                        "type": "integer",
                        "format": "integer",
                        "example": "99eb9c87-79ab-491d-9bb7-a29e77e14e79"
                    },
                    "name": {
                        "title": "name",
                        "description": "name",
                        "type": "string",
                        "format": "string",
                        "example": "Magnam totam illum."
                    },
                    "starting_date": {
                        "title": "starting date",
                        "description": "starting date",
                        "type": "string",
                        "format": "date",
                        "example": "2023-08-18"
                    },
                    "ending_date": {
                        "title": "ending date",
                        "description": "ending date",
                        "type": "string",
                        "format": "date",
                        "example": "2023-08-23"
                    },
                    "price": {
                        "title": "ending date",
                        "description": "ending date",
                        "type": "number",
                        "format": "float",
                        "example": "154.60"
                    }
                },
                "type": "object"
            },
...
            "TourResource": {
                "title": "TourResource",
                "description": "Tour resource",
                "properties": {
                    "data": {
                        "title": "data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Tour"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "TourResource"
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}